% Generated by build script
\name{rdap_autnum}
\alias{rdap_autnum}
\title{RDAP autonomous system number lookup}
\usage{
rdap_autnum()
}
\arguments{
\item{asn}{Autonomous System Number (e.g. 15169 or "15169")}
}
\value{tibble: handle, name, type, start_autnum, end_autnum,}
\description{RDAP autonomous system number lookup}
