% Generated by build script
\name{rdap_ip}
\alias{rdap_ip}
\title{RDAP IP address/network lookup}
\usage{
rdap_ip()
}
\arguments{
\item{ip}{IP address or CIDR block (e.g. "8.8.8.8", "192.168.1.0/24")}
}
\value{tibble: handle, name, type, start_address, end_address,}
\description{RDAP IP address/network lookup}
